perf: add util module
